+++ ChangeLog.md
@@ -16,6 +16,7 @@
 * Prise en compte correcte des malus d'armure et de casques pour les compétences sur la fiche.
 
 ### Autres améliorations
+* Ajout d'options pour des sons de réussite et échec critiques d'attaques par défaut.
 * Ajout de modificateurs d'attaque +n pour les armes magiques.
 * La durée restante des effets est affichée pour le MJ dans le statut des persos.
 * !cof-effet-temp ne fait entrer en combat que si au moins une des cibles est effectivement affectée.